backend/requirements.txt
Flask==3.1.2
flask-pymongo==2.3.0
pymongo==4.6.0
dnspython==2.4.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
